requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0
google-cloud-secret-manager==2.21.1
httpx==0.28.1
python-dateutil==2.9.0
pydantic==2.10.3
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.24.0
